Conflicts:
	modules/imgcodecs/include/opencv2/imgcodecs.hpp

+++ modules/superres/src/btv_l1.cpp
@@ -185,7 +185,7 @@ namespace
                ocl::KernelArg::WriteOnlyNoSize(forwardMap),
                ocl::KernelArg::WriteOnly(backwardMap));
 
-        size_t globalsize[2] = { size.width, size.height };
+        size_t globalsize[2] = { (size_t)size.width, (size_t)size.height };
         return k.run(2, globalsize, NULL, false);
     }
 
@@ -258,7 +258,7 @@ namespace
         k.args(ocl::KernelArg::ReadOnly(src),
                ocl::KernelArg::ReadWriteNoSize(dst), scale);
 
-        size_t globalsize[2] = { src.cols, src.rows };
+        size_t globalsize[2] = { (size_t)src.cols, (size_t)src.rows };
         return k.run(2, globalsize, NULL, false);
     }
 
@@ -316,7 +316,7 @@ namespace
                ocl::KernelArg::ReadOnlyNoSize(src2),
                ocl::KernelArg::WriteOnly(dst, cn));
 
-        size_t globalsize[2] = { src1.cols * cn, src1.rows };
+        size_t globalsize[2] = { (size_t)src1.cols * cn, (size_t)src1.rows };
         return k.run(2, globalsize, NULL, false);
     }
 
@@ -436,7 +436,7 @@ namespace
         k.args(ocl::KernelArg::ReadOnlyNoSize(src), ocl::KernelArg::WriteOnly(dst),
               ksize, ocl::KernelArg::PtrReadOnly(ubtvWeights));
 
-        size_t globalsize[2] = { src.cols, src.rows };
+        size_t globalsize[2] = { (size_t)src.cols, (size_t)src.rows };
         return k.run(2, globalsize, NULL, false);
     }
