+++ modules/video/src/bgfg_gaussmix2.cpp
@@ -787,7 +787,7 @@ bool BackgroundSubtractorMOG2Impl::ocl_apply(InputArray _image, OutputArray _fgm
     if (bShadowDetection)
         kernel_apply.set(idxArg, nShadowDetection);
 
-    size_t globalsize[] = {frame.cols, frame.rows, 1};
+    size_t globalsize[] = {(size_t)frame.cols, (size_t)frame.rows, 1};
     return kernel_apply.run(2, globalsize, NULL, true);
 }
 
@@ -805,7 +805,7 @@ bool BackgroundSubtractorMOG2Impl::ocl_getBackgroundImage(OutputArray _backgroun
     idxArg = kernel_getBg.set(idxArg, ocl::KernelArg::WriteOnly(dst));
     kernel_getBg.set(idxArg, backgroundRatio);
 
-    size_t globalsize[2] = {u_bgmodelUsedModes.cols, u_bgmodelUsedModes.rows};
+    size_t globalsize[2] = {(size_t)u_bgmodelUsedModes.cols, (size_t)u_bgmodelUsedModes.rows};
 
     return kernel_getBg.run(2, globalsize, NULL, false);
 }
